Upgrade cosmiconfig and semantic-release to remove dependency on yaml 1.10.0.
That version of yarn is flagged by yarn audit, although I believe the vulnerable code path isn't reachable in practice in either dependency.

Since yarn audit doesn't have an option for ignoring particular security warnings, it's only possible to use it for effective alerts if the check is generally green.